pyproject.toml
[project]
name = "timetabling"
version = "0.1.0"
description = "School timetabling system using Hybrid CP + Local Search"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ortools>=9.11",
    "sqlalchemy>=2.0",
    "pymysql>=1.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "cryptography>=42.0",
]

[project.scripts]
timetabling = "timetabling.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/timetabling"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
